side pannel was broken down into multiple widgets
